refactor: project service - move out of commands folder
